% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setapikey.R
\name{api_key}
\alias{api_key}
\title{Get or set API_KEY value}
\usage{
api_key(force = FALSE)
}
\arguments{
\item{force}{A boolean to force set new API key for current environment}
}
\value{
A vector containing the API key
}
\description{
This package will need a Weather API key residing in the environment
variable \code{API_KEY}. Set it in the \code{.Renviron} file in the home directory.
}
