% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_astronomy.R
\name{data_request}
\alias{data_request}
\title{get the astronomy data of specific date time and city from weatherapi.com}
\usage{
data_request(query)
}
\arguments{
\item{query}{A list contain information of city, date, and api_key, represented by q, dt, key, repectively.}
}
\value{
The http get response containing astronomy information.
}
\description{
get the astronomy data of specific date time and city from weatherapi.com
}
\examples{
data_request(list(key="abc", q="London", dt="2021-01-01"))

}
