% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_weather.R
\name{get_current_weather}
\alias{get_current_weather}
\title{Get current weather information}
\usage{
get_current_weather(location, air_quality = "yes")
}
\arguments{
\item{location}{A vector of location}

\item{air_quality}{A string of "yes" or "no" to show air quality
information. Default is show air quality information ("yes").}
}
\value{
A data frame of current weather information (see vignette for details)
}
\description{
Get the current real-time weather and air-quality information
}
\examples{
get_current_weather("Kelowna")
get_current_weather(c("Kelowna", "Vancouver"), "no")
}
