% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_history_daily_weather.R
\name{get_history_daily_weather}
\alias{get_history_daily_weather}
\title{Get history weather information for a specific day}
\usage{
get_history_daily_weather(key, q, dt)
}
\arguments{
\item{key}{A string containing the API key}

\item{q}{A string containing the location, which can be US Zipcode, UK Postcode, Canada Postalcode, IP address, Latitude/Longitude (decimal degree) or city name.}

\item{dt}{A string containing the date for query, which should be within the last 7 days.}
}
\value{
A data frame containing the daily weather information for the day.
}
\description{
Get history weather information for a specific day
}
\examples{
get_history_daily_weather("YourAPIKeyHere","London","2022-02-12")
}
