% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_zone.R
\name{get_time_zone}
\alias{get_time_zone}
\title{Get time zone information}
\usage{
get_time_zone(location)
}
\arguments{
\item{location}{A vector of location}
}
\value{
A data frame of local time-zone information (see vignette for details)
}
\description{
Get the local time-zone information for a location
}
\examples{
get_time_zone("Kelowna")
get_time_zone(c("Kelowna", "Vancouver"))
}
